Fix time zone issue.
When creating a new date, node gives you a date in your time zone. My laptop is in Europe/london whereas the teamcity server is UTC. This lead to my test failing in teamcity only.

The solution is to create all the date in UTC, and ensuring all date operation are done using their UTC equivalent.