Parse the `app_account_token` from Apple webhook events (if available) (#1403)
When Apple make a request to the webhook endpoint they can optionally provide a
app_account_token UUID (nested under the unified_receipt data). We'll be making
use of this field for Feast (and eventually for other apps) so start parsing it
from the payload. We don't do anything with it yet, that will happen in future
PRs.

Note: This PR includes moving some code out of the pubsub/apple.ts handler file
and into pubsub/apple-common.ts to reflect the fact that they're shared by the
main handler as well as the Feast handler.