Add Feast event data to events Dynamo table (#1402)
When the webhook endpoint for Feast is called, add the event data to the Dynamo
events table (the same table that other app events are added to).

Includes a little refactor around the definition of handler/buildHandler. I
think this makes it slightly clearer that the handler is 100% defined in terms
of buildHandler, which does the dependency injection. It also means types
aren't repeated across the two definitions.