Feast product name should be `FeastInAppPurchase` in soft-opt-in queue (#1414)
The soft-opt-in consents for Feast IAPs are different from other IAPs. In the
soft-opt-in-consent-setter lambda the config has been updated to specify the
new consents, but that lambda needs to be able to distinguish Feast IAPs from
others. So in this PR we start providing a different productName
(FeastInAppPurchase) for Feast events.