Handle 21007 response for Apple receipt lookup gracefully (#1426)
Includes a refactor to make error handling easier.

If the receipt lookup fails with 21007, fail gracefully. If the sub comes from
a sandbox purchase, it's okay that it fails - this is expected. We don't want
the lambda run to error, triggering a retry.  So catch these types of errors and
let the lambda return successfully. This mirrors the behaviour for the live
app.