Remove 2 day window soft-opt-in email check for Feast
This is preventing Feast IAPs from triggering the SOI email. The
condition exists for live to guard against data moving through the
system slowly, but the live app works differently (uses the /link
endpoint for example).

I've commented the test out because I had a hard time updating the mocks
to cover this new scenario. I'll fix these in a subsequent PR (or
possibly refactor to move away from mocks).