More explicit validation for Google pubsub payloads
No more casting. This way if it's not a payload we can handle we fail
early instead of throwing a cryptic runtime error later on.

We know that currently we don't handle other types of notifications,
other than subscription notifications. We see lots of errors resulting
from the lambda being invoked by a voided purchase notification in the
logs. This doesn't really change here - but these will now be explicitly
rejected as we'll fail to parse them. The next step will be to handle
these notifications, but do nothing (returning a 2xx response).