Feast subscription update lambda: write the subscription record, even if app account token is missing (#1463)
We've identified some cases where a receipt for Feast may not contain an
appAccountToken, which is how we link to an identity ID. Handling these
correctly and ensuring they are linked with an identity ID is something we're
looking at separately. In the meantime, the lack of appAccountToken currently
causes us to fail to write the subscription record AND the link record. But we
shouldn't let it prevent us from writing the subscription. It's useful data,
and makes it easier for us to identify the records which don't have a link to
an identity ID.