Address Android pubsub errors (#1465)
In production we see a large volume of errors due to the payload sent to the
pubsub endpoint by Google not being what we expect. We expect to always receive
a [subscription notification] and we (before this PR) cast to a type which
matches this.

After adding some logging in #1461 I can see that we often receive a voided
[purchase notification]. So far I've not been able to get clarity on what we
should be doing with these, so let's change the behaviour to fail gracefully
(and log) until we've figured out what to do. Now when we receive a voided
purchase notification we'll do nothing and return a 200.

I've introduced [zod] to do more explicit validation. I hope to extend this to
other areas of the codebase in future PRs.

[subscription notification]: https://developer.android.com/google/play/billing/rtdn-reference#sub
[purchase notification]: https://developer.android.com/google/play/billing/rtdn-reference#voided-purchase
[zod]: https://zod.dev/