Increase the threshold of this DLQ alarm
We seem to have a small trickle of these which fail (which we have an
ongoing investigation into) which means this alarm is always firing. I'm
concerned that if we pushed out a change which broke _everything_ we
wouldn't know about it, as the alarm is already firing. So increase the
threshold for now.