The link endpoint will now handle Feast IAPs (in some cases)
When a user takes out an IAP using an offer code, the appAccountToken is
not set in the receipt. So we need a different way to link identity IDs
with IAPs. We've decided to use the link endpoint used by the live apps
to facilitate this. It will only be called for promo code purchases.
Other Feast purchases will use the appAccountToken as usual.