Set the maximum number of retry events for the SOI acquisitions lambda
This lambda is triggered by a stream of Dynamo events. We've encountered
a scenario where the user deleted their account before this lambda ran,
which means we're continually failing to get the user's email address
from identity.

If the lambda errors during invocation, we get retries for free:
https://docs.aws.amazon.com/lambda/latest/dg/with-ddb.html#services-dynamodb-errors

The default retry behaviour is to retry indefinitely:
https://docs.aws.amazon.com/lambda/latest/api/API_CreateEventSourceMapping.html
(See MaximumRetryAttempts).

There's no point in retrying forever, so set the retry limit to 10 as a
starter.